Rework rule editor to fix broken text input, slider for scan range, simplify command
Rule fields lived inside a NestedListListEntry cell (list-of-lists, two levels
of nesting) where keyboard text input didn't reliably reach the field, while
click-based controls worked fine at that depth. Each rule is now its own
SubCategory (one level of nesting, matching General's already-working fields)
instead. Cloth Config has no plain button widget (checked directly against
the jar), so add/delete are "toggle, then press Save" actions reconciled in
a new Module.onConfigScreenSaved() hook, called before the single write to
disk and available to any future module needing the same pattern.

Scan range is now an int slider (0-128) instead of a typed double field.

/k8ba dropped back to a single literal — the case-permutation registration
wasn't the least bit different in registration mechanics from the plain
literal, so removing it isolates the command down to Fabric's actual
registration path; added a log line on execution to help diagnose further
if it's still not firing.

+++ src/main/java/com/k8bas/skyblockutility/config/GeneralConfig.java
@@ -4,5 +4,5 @@ public class GeneralConfig {
 	public boolean autoUpdateCheckEnabled = false;
 	public boolean autoUpdateDownloadEnabled = true;
 	/** Blanket cap, in blocks, on how far away mobs are considered for highlighting. 0 = unlimited. */
-	public double mobScanRangeBlocks = 64;
+	public int mobScanRangeBlocks = 64;
 }
